HHH-18489 Test initialization of unowned, lazy one-to-one associations
